Correct and extend AssignNewDcaValues.C
The logic in the previous version of this script was incorrect.
Now, the macro takes DCAxy values of an external file and assigns
fake DCAxy values to heavy-flavor events. Additional leafs are
added to the tree structure, containing the original DCAxy values
for non-HF events and fake DCAxy values for HF events.