Add macro for merging processed MVA outputs into a tree
Process MVA output values from an external file and, depending on
a user-defined cut value, decide whether an event is tagged as
either accepted or not accepted. Subsequently, add the resulting
(boolean) values to an existing tree as a new branch, using a
user-chosen branch name. (Note: The existing ROOT file is NOT
copied in the process, but merely updated.)