Add feature
If no pretrained model is loaded, the program now informs the user
that model training is going to be invoked and waits a couple of
seconds in case the user wants to cancel training (and thus
overwriting of existing training data). If no input is received
in this time, the program proceeds in the usual way.