Add script
Script for tree preparation for loose track cut processing by
the prefilter:
The script takes two trees as inputs and assigns a label in form
of an additional branch to each of them, indicating whether they
originate from the tree with "normal" track cuts or the tree with
"looser" ones. Subsequently, the two trees are merged and the
resulting tree (now also including the label/branch "IsLooseCuts")
is stored as a new ROOT file.