Rewrite script
Rewrite the entire script and change its logic, because the prior
implementation was not able to properly handle the different
directories in the input trees. This issue became apparent only
when running the pairing macro on the merged tree later on (in
form of zero-only entries in the new branch 'IsLooseCuts').