Implement unique event ID assignment
Coming from the Grid, the data is organized in many trees, each
of them counting event IDs beginning from 1. Therefore, in the
"merged" tree, events with the same ID do not necessarily belong
to the same physical event.
With this commit, a new variable/branch is introduced which
contains event IDs that are unique throughout the (merged) tree.